+++ NewFile.txt
@@ -198,3 +198,23 @@
 123-54-9876,
 111-10-1000,
 108-22-1905,
+246-53-4555,
+135-79-1357,
+107-10-1080,
+219-02-1962,
+123-23-1234,
+123-47-2476,
+321-54-9800,
+685-14-8507,
+344-54-9802,
+190-19-1907,
+321-54-9876,
+098-68-1985,
+555-55-5555,
+777-77-7777,
+123-54-9876,
+111-10-1000,
+108-22-1905,
+321-44-5555,
+659-45-5555,
+365-52-6547,
