+++ configuration.properties
@@ -9,7 +9,7 @@ validPassword=123qw?+.
 employee_id=group8employee
 employee_password=group8employee
 new_Applicant_SSN=333-33-1234
-api_bearer_token=eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJncm91cDhhZG1pbiIsImF1dGgiOiJST0xFX0FETUlOIiwiZXhwIjoxNjA2MjY0MTc0fQ.yWnEag0xG9w0QFfCxpDZmo9aDbX5EX1VZZDzI3_hh2jC3uo2SaFfX-_LPHRkKZY51S8Ki_J9e3VPFAAURKokNw
+api_bearer_token=eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJncm91cDhhZG1pbiIsImF1dGgiOiJST0xFX0FETUlOIiwiZXhwIjoxNjA2MzIzMzgxfQ.-6_rGyP4OrdikkJyV6f_OiZLzjApMj90SzPuyrvBLWS_TlSaTY1GAvF7ORWXX63Lbz8UeK_nEFmwCdtZYupU-A
 filePath_Customer=CustomerData.txt
 api_endpoint_customers=https://www.gmibank.com/api/tp-customers
 api_endpoint_states=https://www.gmibank.com/api/tp-states
