+++ src/test/java/gmibank/stepdefinitions/LoginTestStepdefinitions.java
@@ -12,7 +12,7 @@ public class LoginTestStepdefinitions {
 
     @Given("user is on the login page")
     public void user_is_on_the_login_page() {
-        Driver.getDriver().get(ConfigurationReader.getProperty("loginUrl"));
+        Driver.getDriver().get(ConfigurationReader.getProperty("login_url"));
 
     }
