+++ us20.csv
@@ -1,0 +1,40 @@
+23561 , null , null
+22163 , null , null
+18440 , null , null
+5129 , null , null
+5449 , null , null
+22175 , null , null
+22176 , null , null
+22177 , null , null
+3870 , null , null
+5407 , null , null
+9260 , null , null
+9268 , null , null
+3862 , null , null
+18414 , null , null
+5143 , null , null
+19770 , null , null
+5414 , null , null
+5418 , null , null
+6476 , null , null
+6478 , null , null
+23561 , null , null
+22163 , null , null
+18440 , null , null
+5129 , null , null
+5449 , null , null
+22175 , null , null
+22176 , null , null
+22177 , null , null
+3870 , null , null
+5407 , null , null
+9260 , null , null
+9268 , null , null
+3862 , null , null
+18414 , null , null
+5143 , null , null
+19770 , null , null
+5414 , null , null
+5418 , null , null
+6476 , null , null
+6478 , null , null
